requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-hint==0.3.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
inflection==0.5.1
kombu==5.3.7
Markdown==3.6
multidict==6.0.5
packaging==24.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4

mmh3~=4.1.0